pyproject.toml
[project]
name = "asterley-bros"
version = "0.1.0"
description = "AI-powered lead generation and outreach system for Asterley Bros craft spirits"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "camoufox[geoip]>=0.4",
    "playwright>=1.40",
    "firebase-admin>=6.0",
    "google-genai>=1.0",
    "anthropic>=0.40",
    "resend>=2.0",
    "streamlit>=1.30",
    "fastapi>=0.115",
    "uvicorn[standard]>=0.34",
    "python-dotenv>=1.0",
    "pydantic>=2.0",
    "pydantic-settings>=2.0",
    "tenacity>=8.0",
    "structlog>=24.0",
    "pyyaml>=6.0",
    "httpx>=0.27",
    "beautifulsoup4>=4.12",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0",
    "pytest-asyncio>=0.23",
]
